Design+deploy the facebook opengraph action UI with a cool button.
We don't actually need to delete before we create an action because
we know if the action already exists. If it exists, we're done. If it
doesn't exist, we can delete actions for the other verbs asyncrhonously
since there's no mutual exclusion rule.

Remove the old, simpler FB like widget.

Remove the need for #fb-staging in the URL.