the change to STATIC_URL broke bill widgets in 9016b0938bf759a542bc10c9947278da956e0bf7
The bill widget HTML template seems to be the only template that doesn't
extend master.html. In master.html we explicitly define STATIC_URL because
Django's error views (500/400 pages) load the template without any context
variables. That seemed to make all of the other pages on the site work by
accident.

The fix is to add the staticfiles template context processor to settings.py
so that STATIC_URL gets set the right way. We still need the fix in
master.html because this is what the 500 pages don't load.