overhaul indexing of MoC roles and MoC search
On the MoC search/list pages, the faceting of historical members was totally broken because the facets were based on the most recent role of each person. So when looking for everyone that served in the house, for instance, it would only show members whose most recent role was in the house - not anyone that later served in the senate.

Now we separate current MoC search (which facets based in current roles) from historical MoC search (which facets based on multi-valued fields that has info on all of a MoC's roles).