bill prognosis was not being cached because the value was over memcached limits
So every request for a prognosis, especially in bill search results, was hitting the hard disk
for each bill.

To stay within the 1 MB value limit, we only cache the prognosis number for each bill,
not the explanatory text. The explanatory text is only used on bill pages, so we keep
going to disk for those pages.