Make "relevance" the default search when searching for text
This replaces the previous bahavior which relied on Solr "boosting"
text-relevant matches, after noticing that Solr sometimes badly fails
to uprank highly-relevant text matches.

The example is a search for "john mccain national defense authorization"
which failed to put the bills with the corresponding titles in the first
several dozen results. The new default of text sorts ordered by relevance
fixes this unintuitive behavior.