Show middlenames and suffixes when it's needed to distinguish the person
Show middle names and suffixes (but not both) when it's necessary to distinguish people from other people with the same first and last name and have served in the same role. Suffixes are included only when showing first names before last names, which was generally the logic before, but some contexts that didn't include suffixes may now get it and contexts that had it will lose it when it wouldn't be needed to distinguish the person.